[#189] style: Mock 객체가 아닌 Stub이므로 이름 변경하고 파일 분리
