+++ iOS/Todo/Todo/Network/NetworkManager.swift
@@ -63,77 +63,3 @@ struct NetworkManager: NetworkManagable {
         }.resume()
     }
 }
-
-struct MockColumnsSuccessStub: NetworkManagable {
-    func requestResource(from urlString: String, method: HTTPMethod, body: Data?, format: String?, headers: [String]?,
-                         resultHandler: @escaping (Data?, URLResponse?, Error?) -> ()) throws {
-        resultHandler(StubJsonData.successColumnsResponseStub, nil, nil)
-    }
-}
-
-struct MockCardDeleteSuccessStub: NetworkManagable {
-    func requestResource(from urlString: String, method: HTTPMethod, body: Data?, format: String?, headers: [String]?,
-                         resultHandler: @escaping (Data?, URLResponse?, Error?) -> ()) throws {
-        resultHandler(StubJsonData.successDeleteResponseStub, nil, nil)
-    }
-}
-
-struct MockCardCreateSuccessStub: NetworkManagable {
-    func requestResource(from urlString: String, method: HTTPMethod, body: Data?, format: String?, headers: [String]?, resultHandler: @escaping (Data?, URLResponse?, Error?) -> ()) throws {
-        resultHandler(StubJsonData.successCreateResponseStub, nil, nil)
-    }
-}
-
-struct MockCardEditSuccessStub: NetworkManagable {
-    func requestResource(from urlString: String, method: HTTPMethod, body: Data?, format: String?, headers: [String]?, resultHandler: @escaping (Data?, URLResponse?, Error?) -> ()) throws {
-        resultHandler(StubJsonData.successEditingResponseStub, nil, nil)
-    }
-}
-
-struct MockCardMoveSameColumnsSuccessStub: NetworkManagable {
-    func requestResource(from urlString: String, method: HTTPMethod, body: Data?, format: String?, headers: [String]?, resultHandler: @escaping (Data?, URLResponse?, Error?) -> ()) throws {
-        resultHandler(StubJsonData.successMoveSameColumnsResponseStub, nil, nil)
-    }
-}
-
-struct MockCardMoveDifferentColumnsSuccessStub: NetworkManagable {
-    func requestResource(from urlString: String, method: HTTPMethod, body: Data?, format: String?, headers: [String]?, resultHandler: @escaping (Data?, URLResponse?, Error?) -> ()) throws {
-        resultHandler(StubJsonData.successMoveDifferentColumnsResponseStub, nil, nil)
-    }
-}
-
-struct MockLogsSuccessStub: NetworkManagable {
-    func requestResource(from urlString: String, method: HTTPMethod, body: Data?, format: String?, headers: [String]?,
-                         resultHandler: @escaping (Data?, URLResponse?, Error?) -> ()) throws {
-        resultHandler(StubJsonData.successLogsResponseStub, nil, nil)
-    }
-}
-
-struct MockLogSuccessStub: NetworkManagable {
-    func requestResource(from urlString: String, method: HTTPMethod, body: Data?, format: String?, headers: [String]?,
-                         resultHandler: @escaping (Data?, URLResponse?, Error?) -> ()) throws {
-        resultHandler(StubJsonData.successLogResponseStub, nil, nil)
-    }
-}
-
-struct LoginSuccessStub: NetworkManagable {
-    func requestResource(from urlString: String, method: HTTPMethod, body: Data?, format: String?, headers: [String]?, resultHandler: @escaping (Data?, URLResponse?, Error?) -> ()) throws {
-        resultHandler(StubJsonData.successLoginResponse, LoginHTTPURLResponseStub(), nil)
-    }
-}
-
-final class LoginHTTPURLResponseStub: HTTPURLResponse {
-    init() {
-        super.init(
-            url: URL(string: LoginUseCase.EndPoints.login)!,
-            statusCode: 200,
-            httpVersion: nil,
-            headerFields: ["Authorization": "fakeToken"]
-            )!
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
