+++ iOS/Todo/Todo/Network/Stubs/CardMoveSameColumnsSuccessStub.swift
@@ -1,0 +1,15 @@
+//
+//  CardMoveSameColumnsSuccessStub.swift
+//  Todo
+//
+//  Created by kimdo2297 on 2020/07/02.
+//  Copyright © 2020 Jason. All rights reserved.
+//
+
+import Foundation
+
+struct CardMoveSameColumnsSuccessStub: NetworkManagable {
+    func requestResource(from urlString: String, method: HTTPMethod, body: Data?, format: String?, headers: [String]?, resultHandler: @escaping (Data?, URLResponse?, Error?) -> ()) throws {
+        resultHandler(StubJsonData.successMoveSameColumnsResponseStub, nil, nil)
+    }
+}
